refactor to use free functions for deploy
